+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -321,7 +321,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KSW2KQHKYC2Z24XE4F9TXCDR"
+        "01KSYJC41M2D07PWE7VKCCDDXE"
     )
     r"""The unique identifier of the data source"""