+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KSW2KQHF1M4CX614D57WSD3Q"
+        "01KSYJC41FZ59Z43F9PYWZZH7W"
     )
     r"""The unique identifier of the data source"""