+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasetop.py
@@ -81,7 +81,7 @@ class RetrieveDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-05-30T09:13:59.913Z")
+    updated: Optional[datetime] = parse_datetime("2026-05-31T08:27:57.446Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")