+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KSW2KQHH683KTC01HQRW4JZH"
+        "01KSYJC41J28PQJS4ZDAM0YAEQ"
     )
     r"""The unique identifier of the data source"""