+++ packages/orq-rc/RELEASES.md
@@ -8319,3 +8319,13 @@ Based on:
 - [python v4.10.0-rc.58] packages/orq-rc
 ### Releases
 - [PyPI v4.10.0-rc.58] https://pypi.org/project/orq-ai-sdk/4.10.0-rc.58 - packages/orq-rc
+
+## 2026-05-31 08:30:45
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.768.2 (2.889.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.10.0-rc.59] packages/orq-rc
+### Releases
+- [PyPI v4.10.0-rc.59] https://pypi.org/project/orq-ai-sdk/4.10.0-rc.59 - packages/orq-rc